积分

+++ src/main/java/com/pepper/project/cm/activity/controller/ActivityController.java
@@ -91,7 +91,7 @@ public class ActivityController extends BaseController {
      * 修改活动
      */
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable("id") Integer id, ModelMap mmap)
+    public String edit(@PathVariable("id") Long id, ModelMap mmap)
     {
 //        List<Area> areas = areaService.selectAreaList(new Area());
 //        mmap.put("areas",areas);
@@ -129,7 +129,7 @@ public class ActivityController extends BaseController {
      */
     @RequiresPermissions("cm:activity:list")
     @GetMapping("/detail/{id}")
-    public String detail(@PathVariable("id") Integer id, ModelMap mmap)
+    public String detail(@PathVariable("id") Long id, ModelMap mmap)
     {
         String merchantFlag = getSysUser().getMerchantFlag();
         Activity activity = new Activity();
@@ -160,7 +160,7 @@ public class ActivityController extends BaseController {
     @ApiOperation("社区活动详细")
     @PostMapping("/getDetail")
     @ResponseBody
-    public Object getDetail(Integer id)
+    public Object getDetail(Long id)
     {
         if (id != 0)
         {
